minor refactoring of place stuff
