+++ src/main/java/assets/Objects/Pixel.java
@@ -3,7 +3,6 @@ package assets.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import services.database.DBHandlerPlace;
-import services.place.PlaceData;
 
 import java.awt.*;
