+++ src/main/java/commandHandling/commands/placeCommands/PlaceVerify.java
@@ -7,7 +7,7 @@ import net.dv8tion.jda.api.EmbedBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import services.database.DBHandlerConfig;
-import services.place.PlaceData;
+import assets.Objects.PlaceData;
 
 import java.util.List;
