+++ src/main/java/services/listeners/PlaceListener.java
@@ -3,7 +3,7 @@ package services.listeners;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
-import services.place.PlaceData;
+import assets.Objects.PlaceData;
 
 public class PlaceListener extends ListenerAdapter {
     @Override
